Totally reworked how app is configured and works
The initial implementation was too opinionated limiting which claims and parameters can be specified.  After working with some real implementations I realized that this needed to be totally configurable.  This new design is extremely flexible and not opinionated allowing the JWT and OAuth request to be customized in the configruation allowing far more use cases to be supported.